+++ R/get_captions.R
@@ -1,16 +1,16 @@
-#' Get Captions of a Video
+#' Get Particular Caption Track
 #' 
-#' A few of Youtube videos have caption tracks available from an older 
-#' YouTube API. If that caption track is available, this function returns that,
-#' Or it returns caption track specified by id resource. Check \code{\link{list_caption_tracks}} for more
-#' information.
+#' For getting captions from the v3 API, you must specify the id resource. 
+#' Check \code{\link{list_caption_tracks}} for more information.
 #' 
-#' @param video_id ID of the video whose captions are requested. Required. No default.
-#' @param lang  Language of the caption; required; default is \code{"en"} (English)
-#' @param id    String. id of the caption track that is being retrieved
+#' @param id   String. Required. id of the caption track that is being retrieved
+#' @param lang Optional. Default is \code{en}.
+#' @param format Optional. Default is \code{sbv}.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @return String. 
+#' 
+#' @references \url{https://developers.google.com/youtube/v3/docs/captions/download}
 #' 
 #' @export
 #'  
@@ -19,32 +19,17 @@
 #' 
 #' # Set API token via yt_oauth() first
 #' 
-#' get_captions(video_id="yJXTXN4xrI8")
-#' get_captions(id="y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
+#' get_captions(id = "y3ElXcEME3lSISz6izkWVT5GvxjPu8pA")
 #' }
 
-get_captions <- function (video_id=NULL, lang = "en", id = NULL, ...) {
+get_captions <- function (id = NULL, lang = "en", format = "sbv", ...) {
 
-	if (!is.character(video_id) & !is.character(id)) stop("Must specify a valid video_id or id.")
+  if ( !is.character(id)) {
+    stop("Must specify a valid id.")
+  }
 
-	# Try getting captions directly
-	req <- GET(paste0("http://video.google.com/timedtext?lang=", lang, "&v=", video_id))
-	req <- content(req)
+  querylist <- list(tlang = lang, tfmt = format)
+  raw_res <- tuber_GET(paste0("captions", "/", id), query = querylist, ...)
 
-	# If not try other things
-	if (length(req)==0) {
-		
-		querylist = list(id = id)
-		raw_res <- tuber_GET("captions", query = querylist, ...)
-		
-		if (length(raw_res$items) ==0) { 
-    		warning("No caption tracks available. Likely cause: Incorrect video ID. \n")
-    		return(list())
-    	}
-
-		return(raw_res)
-	}
-
-	req
+  raw_res
 }
-